[Dungeon] Remove scaling from default dungeon
It's too much effort to add new features while maintaining the scaling feature, which was mostly added for fast floor planning. Now that the dungeon is planned on a separate thread, this is not as big a concern. Always running at a scale of one makes many new features feasible going forward.