[Shard] Workaround NPE in server for containers in menus
So, this chest here in the virtual inventory menu causes a NullPointerException in Craftbukkit when it tries to getWorldHandle. This means bind scripts won't fire for inventory_click or inventory_close, breaking inventory menus and allowing anyone to take items from them. Spigot will need to fix the bug for us to use containers in virtual inventories in the future.