Expand custom entity spawning to include spawn data
Spawn data, in this case, refers to data about this particular instance of an entity. For example, we now support 'owner' (for tameables), 'ignore' (a list of ignored players for targeting and the like), and 'source' (a new option for setting things like the source of an explosion). As an object array, spawn data can more easily be expanded upon later without having to modify each use of the procedure, as you see in this commit.