[Timer] Update how rankings work
Any time in a course that beats at least one other player now counts towards points in the total rankings. This increases granularity in rank ups too, to give people a sense of progress.

Additionally, it will always show where you are in the top lists, as well as the player above and below you.